+++ AdNetworkSupport/AdColony/AdColonyInterstitialCustomEvent.h
@@ -11,7 +11,7 @@
 #import <AdColony/AdColony.h>
 
 /*
- * Certified with version 2.2.2 of the AdColony SDK.
+ * Certified with version 2.2.4 of the AdColony SDK.
  *
  * The AdColony SDK does not provide an ad clicked callback. As a result, this custom event will not invoke delegate methods
  * interstitialCustomEventDidReceiveTapEvent: and interstitialCustomEventWillLeaveApplication: