+++ AdNetworkSupport/Greystripe/GreystripeInterstitialCustomEvent.h
@@ -11,7 +11,7 @@
 #import "GSAdDelegate.h"
 
 /*
- * Certified with version 4.2.2 of the Greystripe SDK.
+ * Certified with version 4.2.3 of the Greystripe SDK.
  */
 
 @interface GreystripeInterstitialCustomEvent : MPInterstitialCustomEvent <GSAdDelegate>