+++ MoPubSDK/Internal/Common/MPAdServerURLBuilder.m
@@ -34,6 +34,8 @@ NSString * const kMoPubInterfaceOrientationLandscape = @"l";
 + (NSString *)queryParameterForISOCountryCode;
 + (NSString *)queryParameterForMobileNetworkCode;
 + (NSString *)queryParameterForMobileCountryCode;
++ (NSString *)queryParameterForDeviceName;
++ (NSString *)queryParameterForTwitterAvailability;
 + (BOOL)advertisingTrackingEnabled;
 
 @end
@@ -68,6 +70,7 @@ NSString * const kMoPubInterfaceOrientationLandscape = @"l";
     URLString = [URLString stringByAppendingString:[self queryParameterForMobileNetworkCode]];
     URLString = [URLString stringByAppendingString:[self queryParameterForMobileCountryCode]];
     URLString = [URLString stringByAppendingString:[self queryParameterForDeviceName]];
+    URLString = [URLString stringByAppendingString:[self queryParameterForTwitterAvailability]];
 
     return [NSURL URLWithString:URLString];
 }
@@ -200,6 +203,19 @@ NSString * const kMoPubInterfaceOrientationLandscape = @"l";
 {
     NSString *deviceName = [[UIDevice currentDevice] hardwareDeviceName];
     return deviceName ? [NSString stringWithFormat:@"&dn=%@", [deviceName URLEncodedString]] : @"";
+}
+
++ (NSString *)queryParameterForTwitterAvailability
+{
+    MPTwitterAvailability twitterAvailability = [[MPInstanceProvider sharedProvider] twitterAvailabilityOnDevice];
+    NSString *queryString = @"";
+
+    if (twitterAvailability)
+    {
+        queryString = [NSString stringWithFormat:@"&ts=%u", twitterAvailability];
+    }
+
+    return queryString;
 }
 
 + (BOOL)advertisingTrackingEnabled