+++ MoPubSDK/Internal/MPInstanceProvider.h
@@ -70,6 +70,12 @@
 
 typedef id(^MPSingletonProviderBlock)();
 
+typedef enum {
+    MPTwitterAvailabilityNone = 0,
+    MPTwitterAvailabilityApp = 1 << 0,
+    MPTwitterAvailabilityNative = 1 << 1,
+} MPTwitterAvailability;
+
 @interface MPInstanceProvider : NSObject
 
 + (MPInstanceProvider *)sharedProvider;
@@ -137,5 +143,8 @@ typedef id(^MPSingletonProviderBlock)();
 - (NSDictionary *)sharedCarrierInfo;
 
 - (MPTimer *)buildMPTimerWithTimeInterval:(NSTimeInterval)seconds target:(id)target selector:(SEL)selector repeats:(BOOL)repeats;
+
+- (MPTwitterAvailability)twitterAvailabilityOnDevice;
+- (void)resetTwitterAppInstallCheck;
 
 @end