+++ Specs/MoPubSDKSpecs/Extensions/TWTweetComposeViewController+MPSpecs.h
@@ -1,0 +1,15 @@
+//
+//  TWTweetComposeViewController+MPSpecs.h
+//  MoPubSDK
+//
+//  Created by Evan Davis on 2/27/14.
+//  Copyright (c) 2014 MoPub. All rights reserved.
+//
+
+#import <Twitter/Twitter.h>
+
+@interface TWTweetComposeViewController (MPSpecs)
+
++ (void)setNativeTwitterAvailable:(BOOL)available;
+
+@end