+++ Specs/MoPubSDKSpecs/Internal/Common/MPAdServerURLBuilderSpec.mm
@@ -2,6 +2,7 @@
 #import "MPConstants.h"
 #import "MPIdentityProvider.h"
 #import "MPGlobal.h"
+#import "TWTweetComposeViewController+MPSpecs.h"
 #import <CoreLocation/CoreLocation.h>
 
 using namespace Cedar::Matchers;
@@ -238,6 +239,51 @@ describe(@"MPAdServerURLBuilder", ^{
                                            location:nil
                                             testing:YES];
         URL.absoluteString should contain([NSString stringWithFormat:@"&dn=%@", [[[UIDevice currentDevice] hardwareDeviceName] URLEncodedString]]);
+    });
+
+    describe(@"Twitter Availability", ^{
+        beforeEach(^{
+            [fakeProvider resetTwitterAppInstallCheck];
+            [[UIApplication sharedApplication] setTwitterInstalled:NO];
+            [TWTweetComposeViewController setNativeTwitterAvailable:NO];
+        });
+
+        it(@"should not pass a query if no twitter is available", ^{
+            URL = [MPAdServerURLBuilder URLWithAdUnitID:@"guy"
+                                               keywords:nil
+                                               location:nil
+                                                testing:YES];
+            URL.absoluteString should_not contain([NSString stringWithFormat:@"&ts="]);
+        });
+
+        it(@"should create query for only having twitter app installed", ^{
+            [[UIApplication sharedApplication] setTwitterInstalled:YES];
+            URL = [MPAdServerURLBuilder URLWithAdUnitID:@"guy"
+                                               keywords:nil
+                                               location:nil
+                                                testing:YES];
+            URL.absoluteString should contain([NSString stringWithFormat:@"&ts=1"]);
+        });
+
+        it(@"should create query for only having native twitter account(s)", ^{
+            [TWTweetComposeViewController setNativeTwitterAvailable:YES];
+            URL = [MPAdServerURLBuilder URLWithAdUnitID:@"guy"
+                                               keywords:nil
+                                               location:nil
+                                                testing:YES];
+            URL.absoluteString should contain([NSString stringWithFormat:@"&ts=2"]);
+
+        });
+
+        it(@"should create query for having both twitter app and native account(s)", ^{
+            [[UIApplication sharedApplication] setTwitterInstalled:YES];
+            [TWTweetComposeViewController setNativeTwitterAvailable:YES];
+            URL = [MPAdServerURLBuilder URLWithAdUnitID:@"guy"
+                                               keywords:nil
+                                               location:nil
+                                                testing:YES];
+            URL.absoluteString should contain([NSString stringWithFormat:@"&ts=3"]);
+        });
     });
 });