+++ Specs/MoPubSDKSpecs/Internal/MPInstanceProviderSpec.mm
@@ -9,6 +9,7 @@
 #import "MPAnalyticsTracker.h"
 #import "MPHTMLInterstitialViewController.h"
 
+
 using namespace Cedar::Matchers;
 using namespace Cedar::Doubles;